requirements.txt
Django==4.2.4
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
